Use originUrl for objects with no frameId.
onHeadersReceived::tabId is broken for embedded objects and has the
value of -1 ("the request isn't related to a tab"). This breaks the
current embedder url filtering logic that tries to use
webNavigation::getFrame with the out-of-range tabId.

This commit adds a workaround to the particular case and uses
originUrl instead of getFrame->parent->url.

Fixes Adobe Flash embedding on https://boards.4chan.org/f/